--- src/main/java/com/lberkholtz/advancedjava/assignmentfive/model/StockData.java
@@ -1,25 +1,0 @@
-package com.lberkholtz.advancedjava.assignmentfive.model;
-
-import java.text.SimpleDateFormat;
-
-/**
- * Abstract Base class for classes that hold Stock data.
- * Provides common code for such classes including date formatting.
- */
-public abstract class StockData {
-
-    /**
-     * Provide a single SimpleDateFormat for consistency
-     * and to avoid duplicated code.
-     */
-    protected SimpleDateFormat simpleDateFormat;
-
-    /**
-     * Base constructor for StockData classes.
-     * Initialize member data that is shared with sub classes.
-     */
-    public StockData() {
-        simpleDateFormat = new SimpleDateFormat("yyyy/MM/dd");
-    }
-
-}
