--- src/main/java/com/lberkholtz/advancedjava/assignmentfive/util/ProgramConfiguration.java
@@ -1,15 +1,0 @@
-package com.lberkholtz.advancedjava.assignmentfive.util;
-
-import java.util.Properties;
-
-/**
- *  A Utility class that provides application properties and
- *  configuration settings.
- */
-public class ProgramConfiguration {
-
-    public static Properties getApplicationProperties() {
-      return null;
-    }
-
-}
