+++ src/main/java/lberkholtz/advancedjava/assignmentfive/services/SimpleStockService.java
@@ -1,0 +1,90 @@
+package lberkholtz.advancedjava.assignmentfive.services;
+
+import lberkholtz.advancedjava.assignmentfive.model.StockQuote;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.List;
+
+/**
+ * An implementation of the StockService that returns hard coded data.
+ */
+public class SimpleStockService implements StockService {
+
+    /**
+     * Return the current price for a share of stock  for the given symbol
+     *
+     * @param symbol the stock symbol of the company you want a quote for.
+     *               e.g. APPL for APPLE
+     * @return a  <CODE>BigDecimal</CODE> instance
+     * @throws StockServiceException if using the service generates an exception.
+     *                               If this happens, trying the service may work, depending on the actual cause of the
+     *                               error.
+     */
+    @Override
+    public StockQuote getQuote(String symbol) {
+        // a dead simple implementation.
+        return new StockQuote(new BigDecimal(100), Calendar.getInstance().getTime(), symbol);
+    }
+
+    /**
+     * Get a historical list of stock quotes for the provide symbol
+     *
+     * @param symbol the stock symbol to search for
+     * @param from   the date of the first stock quote
+     * @param until  the date of the last stock quote
+     * @return a list of StockQuote instances
+     * @throws   StockServiceException if using the service generates an exception.
+     * If this happens, trying the service may work, depending on the actual cause of the
+     * error.
+     */
+    @Override
+    public List<StockQuote> getQuote(String symbol, Calendar from, Calendar until) {
+        // a dead simple implementation.
+        List<StockQuote> stockQuotes = new ArrayList();
+        Date aDay = from.getTime();
+        while (until.after(aDay)) {
+            stockQuotes.add(new StockQuote(new BigDecimal(100), aDay, symbol));
+            from.add(Calendar.DAY_OF_YEAR, 1);
+            aDay = from.getTime();
+        }
+        return stockQuotes;
+    }
+    /**
+     * Get a historical list of stock quotes for the provide symbol
+     *
+     * @param symbol the stock symbol to search for
+     * @param from   the date of the first stock quote
+     * @param until  the date of the last stock quote
+     * @return a list of StockQuote instances
+     * @throws   StockServiceException if using the service generates an exception.
+     * If this happens, trying the service may work, depending on the actual cause of the
+     * error.
+     */
+    @Override
+    public List<StockQuote> getQuote(String symbol, Calendar from, Calendar until, Interval interval) {
+        // a dead simple implementation.
+        List<StockQuote> stockQuotes = new ArrayList();
+        Calendar workingdate = from;
+        while (workingdate.before(until) || workingdate.equals(until)){
+            Date aDay = workingdate.getTime();
+            stockQuotes.add(new StockQuote(new BigDecimal(100), aDay, symbol));
+            if (interval == Interval.HOURLY) {
+                workingdate.add(Calendar.HOUR_OF_DAY, 1);
+            } else if (interval == Interval.DAILY) {
+                workingdate.add(Calendar.DAY_OF_YEAR, 1);
+            } else if (interval == Interval.WEEKLY) {
+                workingdate.add(Calendar.DAY_OF_YEAR, 7);
+            } else if (interval == Interval.MONTHLY) {
+                workingdate.add(Calendar.MONTH, 1);
+            } else // should never need but just in case
+            {
+                workingdate = until;
+            }
+            aDay = from.getTime();
+        }
+        return stockQuotes;
+    }
+}
