+++ src/main/java/lberkholtz/advancedjava/assignmentfive/services/StockServiceException.java
@@ -1,0 +1,39 @@
+package lberkholtz.advancedjava.assignmentfive.services;
+
+/**
+ * Used to signal a problem with the StockService.
+ */
+public class StockServiceException extends Exception {
+
+    /**
+     * Constructs a new exception with the specified detail message.  The
+     * cause is not initialized, and may subsequently be initialized by
+     * a call to {@link #initCause}.
+     *
+     * @param message the detail message. The detail message is saved for
+     *                later retrieval by the {@link #getMessage()} method.
+     */
+    public StockServiceException(String message) {
+
+        super(message);
+    }
+
+    /**
+     * Constructs a new exception with the specified detail message and
+     * cause.  <p>Note that the detail message associated with
+     * {@code cause} is <i>not</i> automatically incorporated in
+     * this exception's detail message.
+     *
+     * @param message the detail message (which is saved for later retrieval
+     *                by the {@link #getMessage()} method).
+     * @param cause   the cause (which is saved for later retrieval by the
+     *                {@link #getCause()} method).  (A <tt>null</tt> value is
+     *                permitted, and indicates that the cause is nonexistent or
+     *                unknown.)
+     * @since 1.4
+     */
+    public StockServiceException(String message, Throwable cause) {
+
+        super(message, cause);
+    }
+}
