--- src/test/java/com/lberkholtz/advancedjava/assignmentfive/services/StockServiceFactoryTest.java
@@ -1,17 +1,0 @@
-package com.lberkholtz.advancedjava.assignmentfive.services;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertNotNull;
-
-/**
- * JUnit test for <CODE>StockServiceFactory</CODE>
- */
-public class StockServiceFactoryTest {
-
-    @Test
-    public void testGetInstance() {
-        StockService stockService = StockServiceFactory.getInstance();
-        assertNotNull(stockService);
-    }
-}
